requirements.txt
attrs==23.1.0
Automat==22.10.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
crochet==2.0.0
cryptography==41.0.1
cssselect==1.2.0
filelock==3.12.2
Flask==2.3.2
greenlet==2.0.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.2
MarkupSafe==2.1.3
packaging==23.1
parsel==1.8.1
playwright==1.35.0
Protego==0.2.1
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyee==9.0.4
pyOpenSSL==23.2.0
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
Scrapy==2.9.0
scrapy-playwright==0.0.26
scrapy-user-agents==0.1.1
service-identity==23.1.0
six==1.16.0
tldextract==3.4.4
Twisted==22.10.0
typing_extensions==4.7.0
ua-parser==0.16.1
urllib3==2.0.3
user-agents==2.2.0
w3lib==2.1.1
Werkzeug==2.3.6
wrapt==1.15.0
zope.interface==6.0
